refactor: ♻️ 一部のディレクトリ構造を修正
